refactor : save settings in main process
